+++ packages/graph-explorer/src/modules/CreateConnection/CreateConnection.tsx
@@ -43,6 +43,7 @@ type ConnectionForm = {
   awsAuthEnabled?: boolean;
   serviceType?: NeptuneServiceType;
   awsRegion?: string;
+  awsAssumeRoleArn?: string;
   fetchTimeoutEnabled: boolean;
   fetchTimeoutMs?: number;
   nodeExpansionLimitEnabled: boolean;
@@ -72,6 +73,7 @@ function mapToConnection(data: Required<ConnectionForm>): ConnectionConfig {
     awsAuthEnabled: data.awsAuthEnabled,
     serviceType: data.serviceType,
     awsRegion: data.awsRegion,
+    awsAssumeRoleArn: data.awsAssumeRoleArn,
     fetchTimeoutMs: data.fetchTimeoutEnabled ? data.fetchTimeoutMs : undefined,
     nodeExpansionLimit: data.nodeExpansionLimitEnabled
       ? data.nodeExpansionLimit
@@ -164,6 +166,7 @@ const CreateConnection = ({
     awsAuthEnabled: initialData?.awsAuthEnabled || false,
     serviceType: initialData?.serviceType || "neptune-db",
     awsRegion: initialData?.awsRegion || "",
+    awsAssumeRoleArn: initialData?.awsAssumeRoleArn || "",
     fetchTimeoutEnabled: initialData?.fetchTimeoutEnabled || false,
     fetchTimeoutMs: initialData?.fetchTimeoutMs,
     nodeExpansionLimitEnabled: initialData?.nodeExpansionLimitEnabled || false,
@@ -333,6 +336,22 @@ const CreateConnection = ({
                 ]}
                 value={form.serviceType}
                 onValueChange={onFormChange("serviceType")}
+              />
+            </FormItem>
+            <FormItem>
+              <Label>
+                AWS Assume Role ARN
+                <InfoTooltip>
+                  ARN of the role that the proxy-server should assume to sign
+                  requests. This is only required if the connector is running
+                  outside of the AWS account that hosts the Neptune resources.
+                </InfoTooltip>
+              </Label>
+              <InputField
+                data-autofocus={true}
+                value={form.awsAssumeRoleArn}
+                onChange={onFormChange("awsAssumeRoleArn")}
+                placeholder="arn:aws:iam::aws-account-no:role/role-name"
               />
             </FormItem>
           </>